added application testing config, simplified test structure
